:lipstick: :: BaseActivity 만들어서 적용
