+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/FillOutInformationActivity.kt
@@ -1,9 +1,5 @@
 package com.sms.presentation.main.ui.fill_out_information
 
-import android.os.Bundle
-import android.view.WindowManager
-import android.widget.Toast
-import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.runtime.mutableStateOf
@@ -14,6 +10,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.sms.presentation.main.ui.base.BaseActivity
 import com.sms.presentation.main.ui.detail_stack_search.DetailStackSearchScreen
 import com.sms.presentation.main.ui.fill_out_information.screen.*
 import com.sms.presentation.main.viewmodel.FillOutViewModel
@@ -21,16 +18,12 @@ import com.sms.presentation.main.viewmodel.SearchDetailStackViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class FillOutInformationActivity : ComponentActivity() {
+class FillOutInformationActivity : BaseActivity() {
 
     private val fillOutViewModel by viewModels<FillOutViewModel>()
     private val searchDetailStackViewModel by viewModels<SearchDetailStackViewModel>()
 
-    private var doubleBackToExitPressedOnce = false
-    private var backPressedTimestamp = 0L
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
+    override fun viewSetting() {
         fillOutViewModel.getMajorList()
 
         setContent {
@@ -112,28 +105,5 @@ class FillOutInformationActivity : ComponentActivity() {
         }
     }
 
-    override fun onBackPressed() {
-        controlTheStackWhenBackPressed()
-    }
-
-    private fun controlTheStackWhenBackPressed() {
-        val currentTime = System.currentTimeMillis()
-        if (doubleBackToExitPressedOnce && currentTime - backPressedTimestamp <= 2000) {
-            finishAffinity()
-        } else {
-            doubleBackToExitPressedOnce = true
-            backPressedTimestamp = currentTime
-            Toast.makeText(this, "한 번 더 누르시면 앱이 종료됩니다.", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    private fun setSoftInputMode(isType: String = "NOTHING") {
-        window.setSoftInputMode(
-            when (isType) {
-                "PAN" -> WindowManager.LayoutParams.SOFT_INPUT_ADJUST_PAN
-                "RESIZE" -> WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE
-                else -> WindowManager.LayoutParams.SOFT_INPUT_ADJUST_NOTHING
-            }
-        )
-    }
+    override fun observeEvent() {}
 }
